Install ec2nvme-nsid in %{_sbindir} rather than /usr/lib/udev
/usr/lib/udev is not referenced in the FHS or in the Fedora packaging
guidelines, and there are no RPM macros that reference it.  It's best
to treat it as an internal directory managed by the systemd related
packages, and install our external tools to standard system
directories.

Moving the ec2nvme-nsid executable to /usr/sbin is also consistent
with the location of the rest of the executables provided by this
package.