Re-order some CI steps
With GitHub's recent change to enforce stronger hash algorithms,
Circle CI was unable to clone the project repository.  By installing
our dependencies earlier in the workflow, we are able to ensure that
the AL2 git and ssh clients are used, which do support the newer
algorithms.

Details in this discussion on the Circle CI forums:
https://discuss.circleci.com/t/discussion-and-resolution-for-error-youre-using-an-rsa-key-with-sha-1-which-is-no-longer-allowed/42572